common_artifacts/requirements/requirements.txt
click==7.1.2
pyyaml==3.11
marshmallow-dataclass==8.3.0
dataclasses==0.6
pandas==1.2.4
numpy>=1.18
scikit-learn==0.24.1